medichron_fastapi/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Settings management
pydantic==2.5.3
pydantic-settings==2.1.0

# QR Code generation
qrcode[pil]==7.4.2
Pillow==10.2.0

# Data validation and processing
python-dateutil==2.8.2
email-validator==2.1.0

# HTTP requests (optional, for external API calls)
httpx==0.26.0

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# CORS
starlette==0.35.1
